Prevent docker_setup.pl from overwriting .env
The webwork db password is set on first run.  The password in .env
must match, or webwork won't start.  Overwriting an existing .env
with a new randomly-generated webwork db password breaks an existing
container. Make docker_setup.pl check for a .env, and print a helpful
message rather than overwriting it if found.

Also, improve script output to be clearer about what it is doing.