Remove redundant OPL setup from container start
Remove the code in docker-entrypoint.sh that clones the OPL from github
and downloads the OPL metadata.  Update Dockerfile and docker_setup.pl
to match.

The docker_setup.pl script in webwork-docker clones the OPL, and
downloads the metadata, then Docker copies all files to the correct
locations (specified in Dockerfile) as part of the container build
process.  Remove the redundant code from docker-entrypoint.sh to speed
container start.  Updating the OPL tables in the db is the only thing
that needs to happen at container start.