Add packages only at container build time
Remove the code in docker-entrypoint.sh that adds packages at
container start based on the environment variable ADD_APT_PACKAGES.

Changing the value of the environment variable invalidates Docker's
cache, which forces a rebuild, so there are no build-time savings to be
had in delaying until container startup. Adding at startup slows the
container start, however, so there is a cost.

Packages should be added in the build stage. Remove ADD_APT_PACKAGES.