Separate extra packages install to speed rebuilds
Because all apt-get additions were in a single RUN command, changing
the list of additional packages via the ADD_BASE_PACKAGES environment
variable in .env invalidated the cache, causing all packages to be
reinstalled at build.  Move the installation of additional packages to
a separate RUN command, so changing the list only reinstalls
the additional packages.